Move useImageSize hook to common
