+++ src/components/card/mobile/Card.tsx
@@ -2,7 +2,11 @@ import React from 'react'
 
 import cn from 'classnames'
 
-import placeholderArt from 'assets/img/imageBlank2x.png'
+import placeholderArt from 'common/assets/image/imageBlank2x.png'
+import {
+  useCollectionCoverArt,
+  useUserProfilePicture
+} from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import {
   ProfilePictureSizes,
@@ -16,10 +20,6 @@ import RepostFavoritesStats, {
   Size
 } from 'components/repost-favorites-stats/RepostFavoritesStats'
 import UserBadges from 'containers/user-badges/UserBadges'
-import {
-  useCollectionCoverArt,
-  useUserProfilePicture
-} from 'hooks/useImageSize'
 
 import styles from './Card.module.css'
