+++ src/components/collection/mobile/CollectionHeader.js
@@ -6,6 +6,7 @@ import cn from 'classnames'
 import Linkify from 'linkifyjs/react'
 import PropTypes from 'prop-types'
 
+import { useCollectionCoverArt } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { Variant } from 'common/models/Collection'
 import { SquareSizes } from 'common/models/ImageSizes'
@@ -16,7 +17,6 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import ActionButtonRow from 'containers/track-page/components/mobile/ActionButtonRow'
 import StatsButtonRow from 'containers/track-page/components/mobile/StatsButtonRow'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useCollectionCoverArt } from 'hooks/useImageSize'
 import { make, useRecord } from 'store/analytics/actions'
 import { isShareToastDisabled } from 'utils/clipboardUtil'
 import { isDarkMode } from 'utils/theme/theme'
