+++ src/components/create-playlist/CreatePlaylistModal.js
@@ -4,11 +4,11 @@ import { Modal, Button, ButtonSize, ButtonType } from '@audius/stems'
 import { debounce } from 'lodash'
 import PropTypes from 'prop-types'
 
+import { useCollectionCoverArt } from 'common/hooks/useImageSize'
 import { SquareSizes } from 'common/models/ImageSizes'
 import Input from 'components/data-entry/Input'
 import TextArea from 'components/data-entry/TextArea'
 import UploadArtwork from 'components/upload/UploadArtwork'
-import { useCollectionCoverArt } from 'hooks/useImageSize'
 import * as schemas from 'schemas'
 import { resizeImage } from 'utils/imageProcessingUtil'
 import zIndex from 'utils/zIndex'
