+++ src/components/general/CoverPhoto.js
@@ -5,11 +5,11 @@ import PropTypes from 'prop-types'
 import Lottie from 'react-lottie'
 
 import loadingSpinner from 'assets/animations/loadingSpinner.json'
-import imageCoverPhotoBlank from 'assets/img/imageCoverPhotoBlank.jpg'
+import imageCoverPhotoBlank from 'common/assets/image/imageCoverPhotoBlank.jpg'
+import { useUserCoverPhoto } from 'common/hooks/useImageSize'
 import { WidthSizes } from 'common/models/ImageSizes'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import ImageSelectionButton from 'components/image-selection/ImageSelectionButton'
-import { useUserCoverPhoto } from 'hooks/useImageSize'
 
 import styles from './CoverPhoto.module.css'
