+++ src/components/remix-settings-modal/RemixSettingsModal.tsx
@@ -3,6 +3,7 @@ import React, { useCallback, useState, useEffect, useRef } from 'react'
 import { Modal, Button, ButtonSize, ButtonType } from '@audius/stems'
 import { debounce } from 'lodash'
 
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes } from 'common/models/ImageSizes'
 import { Track } from 'common/models/Track'
@@ -10,7 +11,6 @@ import { User } from 'common/models/User'
 import Input from 'components/data-entry/Input'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useTrackCoverArt } from 'hooks/useImageSize'
 import { fullTrackPage } from 'utils/route'
 import { withNullGuard } from 'utils/withNullGuard'
