+++ src/components/track/EditTrackModal.js
@@ -4,9 +4,9 @@ import { Modal, Button, ButtonSize, ButtonType } from '@audius/stems'
 import { mapValues } from 'lodash'
 import PropTypes from 'prop-types'
 
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import { SquareSizes } from 'common/models/ImageSizes'
 import FormTile from 'components/data-entry/FormTile'
-import { useTrackCoverArt } from 'hooks/useImageSize'
 import * as schemas from 'schemas'
 import zIndex from 'utils/zIndex'
