+++ src/containers/add-to-playlist/desktop/AddToPlaylistModal.tsx
@@ -6,6 +6,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import SimpleBar from 'simplebar-react'
 
 import { ReactComponent as IconMultiselectAdd } from 'assets/img/iconMultiselectAdd.svg'
+import { useCollectionCoverArt } from 'common/hooks/useImageSize'
 import { CreatePlaylistSource } from 'common/models/Analytics'
 import { Collection } from 'common/models/Collection'
 import { SquareSizes } from 'common/models/ImageSizes'
@@ -19,7 +20,6 @@ import SearchBar from 'components/search-bar/SearchBar'
 import { ToastContext } from 'components/toast/ToastContext'
 import ToastLinkContent from 'components/toast/mobile/ToastLinkContent'
 import { getCollectionId } from 'containers/collection-page/store/selectors'
-import { useCollectionCoverArt } from 'hooks/useImageSize'
 import { newCollectionMetadata } from 'schemas'
 import { AppState } from 'store/types'
 import { playlistPage } from 'utils/route'
