+++ src/containers/artist-recommendations/ArtistRecommendations.tsx
@@ -11,6 +11,7 @@ import { push } from 'connected-react-router'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { ReactComponent as IconClose } from 'assets/img/iconRemove.svg'
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { FollowSource, Name } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { ProfilePictureSizes, SquareSizes } from 'common/models/ImageSizes'
@@ -22,7 +23,6 @@ import FollowButton from 'components/general/FollowButton'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import { MountPlacement } from 'components/types'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { make, useRecord } from 'store/analytics/actions'
 import { AppState } from 'store/types'
 import { useIsMobile } from 'utils/clientUtil'
