+++ src/containers/edit-playlist/mobile/EditPlaylistPage.tsx
@@ -5,7 +5,8 @@ import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
 import { ReactComponent as IconCamera } from 'assets/img/iconCamera.svg'
-import placeholderCoverArt from 'assets/img/imageBlank2x.png'
+import placeholderCoverArt from 'common/assets/image/imageBlank2x.png'
+import { useCollectionCoverArt } from 'common/hooks/useImageSize'
 import { CreatePlaylistSource } from 'common/models/Analytics'
 import { Collection } from 'common/models/Collection'
 import { ID } from 'common/models/Identifiers'
@@ -34,7 +35,6 @@ import TextElement, { Type } from 'containers/nav/mobile/TextElement'
 import { useTemporaryNavContext } from 'containers/nav/store/context'
 import UploadStub from 'containers/profile-page/components/mobile/UploadStub'
 import useHasChangedRoute from 'hooks/useHasChangedRoute'
-import { useCollectionCoverArt } from 'hooks/useImageSize'
 import * as schemas from 'schemas'
 import RandomImage from 'services/RandomImage'
 import { AppState } from 'store/types'
