+++ src/containers/first-upload-modal/FirstUploadModal.tsx
@@ -4,6 +4,7 @@ import { Modal } from '@audius/stems'
 import { connect, useDispatch } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { SquareSizes } from 'common/models/ImageSizes'
 import { getAccountUser } from 'common/store/account/selectors'
@@ -12,7 +13,6 @@ import TwitterButton from 'components/general/TwitterButton'
 import ConnectedMusicConfetti from 'containers/music-confetti/ConnectedMusicConfetti'
 import { show } from 'containers/music-confetti/store/slice'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { useRecord, make } from 'store/analytics/actions'
 import { AppState } from 'store/types'
 import { fullProfilePage } from 'utils/route'
