+++ src/containers/nav/desktop/NavColumn.js
@@ -7,7 +7,8 @@ import { withRouter, NavLink } from 'react-router-dom'
 import SimpleBar from 'simplebar-react'
 import 'simplebar/dist/simplebar.min.css'
 
-import imageProfilePicEmpty from 'assets/img/imageProfilePicEmpty2X.png'
+import imageProfilePicEmpty from 'common/assets/image/imageProfilePicEmpty2X.png'
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { Name, CreatePlaylistSource } from 'common/models/Analytics'
 import { SquareSizes } from 'common/models/ImageSizes'
 import Status from 'common/models/Status'
@@ -43,7 +44,6 @@ import ConnectedProfileCompletionPane from 'containers/profile-progress/Connecte
 import * as signOnActions from 'containers/sign-on/store/actions'
 import { resetState as resetUploadState } from 'containers/upload-page/store/actions'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { make, useRecord } from 'store/analytics/actions'
 import { getAverageColorByTrack } from 'store/application/ui/average-color/slice'
 import { getIsDragging } from 'store/dragndrop/selectors'
