+++ src/containers/notification/components/desktop/UserHeader.tsx
@@ -13,6 +13,7 @@ import { push as pushRoute } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes } from 'common/models/ImageSizes'
 import Status from 'common/models/Status'
@@ -26,7 +27,6 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import Tooltip from 'components/tooltip/Tooltip'
 import { fetchNotificationUsers } from 'containers/notification/store/actions'
 import { getNotificationUserList } from 'containers/notification/store/selectors'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { AppState } from 'store/types'
 import { profilePage } from 'utils/route'
