+++ src/containers/notification/components/mobile/ConnectedUserHeader.tsx
@@ -5,6 +5,7 @@ import { push as pushRoute } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes } from 'common/models/ImageSizes'
 import { User } from 'common/models/User'
@@ -13,7 +14,6 @@ import { formatCount } from 'common/utils/formatUtil'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import { fetchNotificationUsers } from 'containers/notification/store/actions'
 import { getNotificationUserList } from 'containers/notification/store/selectors'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { AppState } from 'store/types'
 
 import styles from './UserHeader.module.css'
