+++ src/containers/now-playing/NowPlaying.tsx
@@ -6,6 +6,7 @@ import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
 import { ReactComponent as IconCaret } from 'assets/img/iconCaretRight.svg'
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import {
   FavoriteSource,
   RepostSource,
@@ -40,7 +41,6 @@ import ShuffleButtonProvider from 'components/play-bar/shuffle-button/ShuffleBut
 import { PlayButtonStatus } from 'components/play-bar/types'
 import { getCastMethod } from 'containers/settings-page/store/selectors'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useTrackCoverArt } from 'hooks/useImageSize'
 import { useFlag } from 'hooks/useRemoteConfig'
 import { HapticFeedbackMessage } from 'services/native-mobile-interface/haptics'
 import { useRecord, make } from 'store/analytics/actions'
