+++ src/containers/play-bar/mobile/PlayBar.tsx
@@ -4,6 +4,7 @@ import cn from 'classnames'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import { FavoriteSource, Name, PlaybackSource } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes } from 'common/models/ImageSizes'
@@ -17,7 +18,6 @@ import FavoriteButton from 'components/general/alt-button/FavoriteButton'
 import PlayButton from 'components/play-bar/PlayButton'
 import TrackingBar from 'components/play-bar/TrackingBar'
 import { PlayButtonStatus } from 'components/play-bar/types'
-import { useTrackCoverArt } from 'hooks/useImageSize'
 import { make, useRecord } from 'store/analytics/actions'
 import {
   getAudio,
