+++ src/containers/search-bar/SearchBar.js
@@ -6,8 +6,8 @@ import { connect } from 'react-redux'
 import { matchPath } from 'react-router'
 import { withRouter } from 'react-router-dom'
 
-import placeholderArt from 'assets/img/imageBlank2x.png'
-import profilePicEmpty from 'assets/img/imageProfilePicEmpty2X.png'
+import placeholderArt from 'common/assets/image/imageBlank2x.png'
+import profilePicEmpty from 'common/assets/image/imageProfilePicEmpty2X.png'
 import { Name } from 'common/models/Analytics'
 import { SquareSizes } from 'common/models/ImageSizes'
 import Bar from 'components/search/SearchBar'
