+++ src/containers/settings-page/components/desktop/VerificationModal.tsx
@@ -6,6 +6,7 @@ import { useDispatch } from 'react-redux'
 
 import { ReactComponent as IconValidationX } from 'assets/img/iconValidationX.svg'
 import { ReactComponent as IconVerified } from 'assets/img/iconVerified.svg'
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { ProfilePictureSizes, SquareSizes } from 'common/models/ImageSizes'
@@ -16,7 +17,6 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import { show as showMusicConfetti } from 'containers/music-confetti/store/slice'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { useRemoteVar } from 'hooks/useRemoteConfig'
 import { useRecord, make, TrackEvent } from 'store/analytics/actions'
 import { profilePage } from 'utils/route'
