+++ src/containers/settings-page/components/mobile/AccountSettingsPage.tsx
@@ -5,6 +5,7 @@ import cn from 'classnames'
 import { debounce } from 'lodash'
 
 import { ReactComponent as IconVerified } from 'assets/img/iconVerified.svg'
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { SquareSizes } from 'common/models/ImageSizes'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
@@ -14,7 +15,6 @@ import Grouping from 'components/groupable-list/Grouping'
 import Row from 'components/groupable-list/Row'
 import { ToastContext } from 'components/toast/ToastContext'
 import SignOutPage from 'containers/nav/mobile/SignOut'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import AudiusBackend from 'services/AudiusBackend'
 import { make, useRecord } from 'store/analytics/actions'
 import {
