+++ src/containers/settings-page/components/mobile/SettingsPage.tsx
@@ -3,6 +3,7 @@ import React, { useContext, useEffect, useCallback } from 'react'
 import cn from 'classnames'
 
 import horizontalLogo from 'assets/img/settingsPageLogo.png'
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
 import { InstagramProfile, TwitterProfile } from 'common/store/account/reducer'
@@ -13,7 +14,6 @@ import GroupableList from 'components/groupable-list/GroupableList'
 import Grouping from 'components/groupable-list/Grouping'
 import Row from 'components/groupable-list/Row'
 import NavContext, { LeftPreset } from 'containers/nav/store/context'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import useScrollToTop from 'hooks/useScrollToTop'
 import Theme from 'models/Theme'
 import { getIsIOS } from 'utils/browser'
