+++ src/containers/settings-page/components/mobile/VerificationPage.tsx
@@ -9,6 +9,7 @@ import {
 } from '@audius/stems'
 import cn from 'classnames'
 
+import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { ProfilePictureSizes, SquareSizes } from 'common/models/ImageSizes'
@@ -19,7 +20,6 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import Page from 'components/general/Page'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useUserProfilePicture } from 'hooks/useImageSize'
 import { useRemoteVar } from 'hooks/useRemoteConfig'
 import { useRecord, make, TrackEvent } from 'store/analytics/actions'
 import { profilePage } from 'utils/route'
