+++ src/containers/sign-on/components/ProfileForm.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types'
 import { Spring } from 'react-spring/renderprops'
 import TwitterLogin from 'react-twitter-auth'
 
-import profilePicEmpty from 'assets/img/imageProfilePicEmpty2X.png'
+import profilePicEmpty from 'common/assets/image/imageProfilePicEmpty2X.png'
 import Input from 'components/data-entry/Input'
 import InstagramAuth from 'components/general/InstagramAuth'
 import ProfilePicture from 'components/general/ProfilePicture'
