+++ src/containers/track-page/components/mobile/TrackHeader.tsx
@@ -4,7 +4,8 @@ import { Button, ButtonType, IconPause, IconPlay } from '@audius/stems'
 import cn from 'classnames'
 import Linkify from 'linkifyjs/react'
 
-import placeholderArt from 'assets/img/imageBlank2x.png'
+import placeholderArt from 'common/assets/image/imageBlank2x.png'
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import { Name } from 'common/models/Analytics'
 import { CID, ID } from 'common/models/Identifiers'
 import { SquareSizes, CoverArtSizes } from 'common/models/ImageSizes'
@@ -19,7 +20,6 @@ import { Size } from 'components/co-sign/types'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import DownloadButtons from 'containers/download-buttons/DownloadButtons'
 import UserBadges from 'containers/user-badges/UserBadges'
-import { useTrackCoverArt } from 'hooks/useImageSize'
 import { useFlag } from 'hooks/useRemoteConfig'
 import { make, useRecord } from 'store/analytics/actions'
 import { isShareToastDisabled } from 'utils/clipboardUtil'
