+++ src/containers/visualizer/VisualizerProvider.tsx
@@ -18,7 +18,7 @@ import { getTheme } from 'store/application/ui/theme/selectors'
 import { shouldShowDark } from 'utils/theme/theme'
 import { make, TrackEvent } from 'store/analytics/actions'
 import { Name } from 'common/models/Analytics'
-import { useTrackCoverArt } from 'hooks/useImageSize'
+import { useTrackCoverArt } from 'common/hooks/useImageSize'
 import { Track } from 'common/models/Track'
 import { SquareSizes } from 'common/models/ImageSizes'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
