+++ src/hooks/useImageSize.test.tsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react'
 
 import { DefaultSizes, SquareSizes } from 'common/models/ImageSizes'
 
-import { useImageSize } from './useImageSize'
+import { useImageSize } from '../common/hooks/useImageSize'
 
 jest.mock('react-redux', () => ({
   useDispatch: () => () => {}
