+++ src/services/AudiusBackend.js
@@ -5,9 +5,9 @@ import * as IdentityAPI from '@audius/libs/src/services/identity/requests'
 import BN from 'bn.js'
 import moment from 'moment-timezone'
 
-import placeholderCoverArt from 'assets/img/imageBlank2x.png'
-import imageCoverPhotoBlank from 'assets/img/imageCoverPhotoBlank.jpg'
-import placeholderProfilePicture from 'assets/img/imageProfilePicEmpty2X.png'
+import placeholderCoverArt from 'common/assets/image/imageBlank2x.png'
+import imageCoverPhotoBlank from 'common/assets/image/imageCoverPhotoBlank.jpg'
+import placeholderProfilePicture from 'common/assets/image/imageProfilePicEmpty2X.png'
 import { Name } from 'common/models/Analytics'
 import FeedFilter from 'common/models/FeedFilter'
 import { DefaultSizes } from 'common/models/ImageSizes'
